services/worker/requirements.txt
celery[redis]==5.4.0
httpx==0.28.1
anthropic==0.42.0
beautifulsoup4==4.12.3
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
